Added experimental 'resource' paths to Store
This is just an idea for a way to dynamically serve css/js etc. See
in-code comments. Not optimized...

Added the standard bootstrap v3.3.7 distribution (css, js and fonts) as
the root of the new "resource dir"

Not sure if we will go this way or not. If we do, need to address and
handle properly setting the content-type, etc, which we don't really do
at all for templates yet