Expand 'view_wrapper' concept with 'type' + strapdown
Added strapdown view_wrapper which uses 'INSERT' instead of 'INCLUDE'
... INSERT is a way to insert the content of a template without
processing it, which means bypass of the post-processor. We need to do
this to be able to override the handling of markdown.

Note that the main reason this feature was added was for better testing
(to be able to compare how strapdown renders). Trying to find balance of
features and simplicity... don't want to make it more complex than is
needed, because there are already 10,000 ways to tackle any problem with
TT and the RapidApp template APIs (and this is actually part of the
challenge)