FIX BUSTED site-one scaffold paths (add temp hack)
When 'default_template_extension' was turned off a while back this
busted the site-one scaffold paths because they relied on adding the
'.html' extension. Need to turn off the global option for the sake of
the posts namespace. So, for now, added a temp hack that falls back to
try '.html' suffix, just for scaffold files.

TODO/FIXME...