site-one: add comment now displays smart "sign in" link
Using the newly added features of RapidApp, if the user is not logged
in, they are presented a sign in link instead of the add form, and
following this add link will log them in, take them back to the post
page and scroll them down to the add comment form.

Also using new log in/out control to keep the user's current page fixed
when logging in and out. Working very nicely, and all via plain web 1.0
techniques (i.e. all new page loads, no Ajax)

Requires RapidApp 1.2111_62