Static scaffold path ConditionalGET + max-age 10 mins
Enabled the ConditionalGET middleware for the default static_path_app
(Plack::App::File) and inlined middleware that sets the cache control to
expire after 10 mins. This is fairly fast, but now that we're returning
304s when the file is not changed, subsequent requests are not very
expensive (but still not totally free, will want to play with this value
later. for now, during dev, sooner is better)