ADDED OFFICIAL 'preview_path' SCAFFOLD OPTION
The scaffold can now declare a 'preview_path' which is a view wrapper
path that should be used to "preview" a post, with the intent being that
it will be able to display the post w/o normal site nav links. This is
now used by the postview page for the (still in-progress) iframe
preview.

The keep-it-simple example site now has a preview template that uses
this. If there is no preview_path defined, it just uses
default_view_path