BEGIN BUILDING SCAFFOLD FOR DISTRIBUTION
Started creating a scaffold 'bootstrap-blog' in the share dir for the
purpose of official distribution with the Rapi::Blog package. Will be
creating a bootstrap script to initialize new sites. This scaffold is
based on 'site-one' example but is being further generalized and made
more dynamic with custom scaffold params.

For now, created a new example site 'bootstrap-blog-default' which
points at the new bootstrap-blog scaffold in the sharedir. This allows
running/testing the scaffold, but separately