Post.open_url_path and add_post_path now support optional mode
The template directive Post.open_url_path can now accept arg of 'direct'
or 'navable' to return the URL via /rapidapp/module/direct or
/rapidapp/module/navable, respectively.

The add_post_path directly only supports 'navable' since direct is not
useful in the context of an add form which relies on opening a new tab
post-save.

For more info on the RapidApp feature this is using see:
https://metacpan.org/pod/RapidApp::Manual::Modules#Direct-Rendering-URLs

Basically this provides a way to open these pages in a nice, fullscreen
mode.

Require RapidApp 1.3000_01