bootstrap-blog: inactive page controls now disabled instead of hidden
Tweaked the 4-item paging control in the bootstrap-blog scaffold to
always show all 4 items, but disable the ones which are N/A (i.e. no
last page if you are already on teh last page, etc). This ensures that
the controls location on teh screen don't move around as items become
available and unavailable.

Note that these page controls are still very plain/rough, which is
intentional to keep the implementation as a useful reference