WORKING SIGNATURE-BASED DB MIGRATIONS (proof-of-concept)
Using the newly added fingerprint feature of DBIx::Class::Schema::Diff,
a working dynamic/signature-based method to run schema migrations on
existing databases is now in place. This works by calculating a
checksum/fingerprint of the state of the current database, and if it
matches a known state, running a migration process.

One migration is now implemented, which is from an old dev version of
the schema (not yet out in the wild). This is just for the
proof-of-concept.

The idea is that these style of migrations can run w/o the need to track
version numbers within the database itself -- it is CAS-based and thus
immutable.

Set dev ver 1.0004_01