NEXT SCHEMA - added 'sections' - WITH AUTO MIGRATION, WIP
Updated the schema and added new table 'sections' with a new section_id
FK in the posts table. This is to become the third taxonomy after tags
and categories with the purpose being to provide a way to associate
posts with a *single* classification. Additionally, sections are
self-referencing so sections can be within sections (i.e. tree)

Also added code to always display the signature of the current database

IN PROGRESS...