ListAPI: new configurable feature _api_params_undef_map
Officially established a mechanism to declare an 'undef' string value
on a per param basis to be used as real 'undef' which cannot be passed
on its own via the string-only URL interface.

We're now using the value 'none' to mean undef for the section_id which
is 'section_id' in list_posts and 'parent_id' in list_sections. Removed
the code which was doing this manually.

The advantage of this implementation is that it can work in both
directions -- both in values supplied by the client, and going back the
other way in the various '_qs' meta attributes returned to the client