Start Rapi::Blog::CatalystApp - common plugin hook
New role/plugin class loaded on the Catalyst application. This allows
hooks into $c. This will be used to add layers to the auth system,
overtop of AuthCore and the standard Catalyst Auth modules which are
automatically loaded, but this will be the spot to hook any custom logic
needed going foreard at the Catalyst plugin layer

IN PROGRESS...