Add column user.disabled with migration + biz-logic
Started new schema migration from v1.02 (fea65238f92786e) which adds the
new column 'disabled' to the user table. This bool flag is now
considered when a user tries to authenticate and if it is set to true,
denies the authentication.

Also regenerated schema classes, and also added the column to the base
DDL