REFACTOR ACCESSSTORE LOGIC TO NEW SCAFFOLD CLASS
Moved/delegated the template scaffold lookup/logic from the AccessStore
class into the new Scaffold class, and theoretically mutliple layered
scaffolds can be accessed, although this has not yet been enabled or
tested. So far, the existing single scaffold has been setup to become
the single configured Scaffold, and everything seems to work

Lots more cleanup and then enable and testing of multiple scaffolds,
which i'm not holding my breath will work as-is, which is the main
reason behind this change.

IN PROGRESS...