WIP MAJOR REFACTORING FOR NEW LAYERED SCAFFOLD CONCEPT
Ongoing, major work on building out new subsystems to allow multiple
layers of scaffolds, which has many implicatons, including concepts of
"claimimg" a given path which can happen for both teh positive (file
exists) and the negative (file doesn't exist, but the scaffold wants to
handle the 404). For this to work properly, the scaffold must also
consider whether or not a given not found path which it might claim is
actually available inthe native RapidApp template namespace. This is now
being handled, which required a new method in RapidApp core.

Split out lots of logic and concepts into multiple object classes, and
some of these may stick around and some may be unfactored. Lots of work
yet to do, but some of the core functionality and structure is now in
place.

Set dev version to 1.0200_03, require dev RapidApp version 1.3102_01

IN PROGRESS...