PRELIMINARY PreauthAction CONTROLLER + STRUCTURE - WIP
Checkpoint commit ...

Began creating a controller and logic/structure for handling pre-auth
action requests. Started 'ActionSession' object class and concept to
facilitate flows where multiple separate events can be handled within
the context of a "session" -- which includes look-up and validation of
an action and support multiple possible flows, such as 1. password reset
2. validates 3. user actually changes password 4. deactivate (since its
single use) but handle different codepoaths such as checking expired,
and possibly allow several follow-up events still within the context of
a single authorized "session"

This is preliminary, but already handles the single-event case of
"valid" as well as a dual event case of expired and then deactivated,
and then "invalid" ... plan to expland on this concept to support
multiple flows, which may need to be expanded across several browser
requests, such as for change password, since the auth needs to be
verified first, then the user needs to supply the new password, etc

IN PROGRESS...