WORKING "Actor" DISPATCH DESIGN FOR PRE-AUTHORIZATIONS
Further fleshed out the "Actor" class design for dispatching/calling
pre-authorizations. This new design made the "ActionSession" concept
redundant and no longer needed, so this class has been unfactored

Actor classes may now handle all details of executing a preauth action,
after it has been validated by the result class and controller. Also
established structure for redirecting the user post-execute, which is
set by the Actor class but still handled in the controller. However, the
Actor object instance has the ctalyst context, so it could optionally
take over the request and respond/detach itself, however, this is not a
planned pattern, b ut it is possible