WIP UNFACTOR/CONSOLIDATE SCAFFOLD ATTRS
Removed and unfactored old/duplicate attrs and configs for scaffolds
which existed before the Scaffold object class refactor, slimming down
both the main Blog.pm builder class and the AccessStore, properly
defering to the Scaffold objects.

Additionally, now that there can be multiple layered Scaffolds, a new,
merged 'scaffold_cfg' attr is being built which provides config params
for all of the Scaffolds at once. Note that there is some overlqap here,
as scaffolds internally use many of their own configs, but also, the
site relies on a single site-wide scaffold config, so this needs to be
provided still. This is one way to do this, the other way might be to
just expose only the first scaffold, although there are pitfalls with
both approaches, as the single scaffold model no longer exists

**THIS REQUIRES MORE TESTING** -- very little test has been done so far,
lots more needs to be done as this change could impact all layers and we
do not yet have good test suite for the new features

IN PROGRESS...