LOGIN TEMPLATE NOW SET FROM SCAFFOLDS
The AuthCore 'login_template' is now set according to the Scaffold
'config login' param. The common underlay now has this param defined,
which means that now all Rapi::Blog instrances will use a login template
specific to it, whether it is in the common underlay or a user scaffold.

The current login template is nearly identical to the one built in to
RapidApp; the only difference so far is the title is hard-coded.

IN PROGRESS...