WIP building out forgot_password client/server working
More progress building out the forgot_password/password_reset workflow,
including tighter/fuller integration with the evolving  'local_info'
API. Added additional client requirement which is to send the query
string "?using=local_info" to give the backend confidence that the
client is expecting and is able to handle responses using this template
var API. The added benefit of this is that now all errors are sent back
using this API when the client has declared it is using it.

Also added real logic steps including looking up the user based on the
supplied username or e-mail. The success response is still
dummy/ploaceholder which says a reset link was sent (but isn't doing
that yet)

IN PROGRESS...