Fix URL auth/redirect + autofocus - REQUIRE RapidApp 1.3102_02
After the copy and prune down of the login page, which was copied from
RapidApp, the username field autofocus, and more importantly, the
auth/redirect RapidApp feature did not work/come over because it relied
on ExtJS which we have currently decided not to load for our
common_underlay login page which is specific to Rapi::Blog, but still
includes the login_form template/block from RapidApp.

Updated RapidApp to make this logic work in more locations, including
without ExtJS (which it never really needed in the first place.

This fix was made as of dev version 1.3102_02, which we now require
(runtime). Specifically, the change was made in this commit:

https://github.com/vanstyn/RapidApp/commit/4b90371f98c279b450108d6d028c1208f0ff09e3

Also enabled autofocus on the forgot_password page, which doesn't and
never did include any RapidApp templates