WIP forgot_password + rework preauth flow
Did a bit of refactoring of the code entry path for Preauth actions.
Decided on a ResultSet method entry point '->request_Actor' which
handles the validation of the request and returning of the Actor object
only if it is authenticated and valid.

Got rid of teh concept of 'not_finished' states after actor execute, and
now strictly enforce a single execute, success or fail, a pre-auth
action can only be executed once. Ifg it fails, a new one would have to
be created. Additionally, the call of execute and the recording of it
now happens within the Actor.

The biggest concept change in this commit is that now we design for the
scenario of any controller logic calling the action, rather than forcing
the built-in controller action. This built-in style via
'remote/preauth_action/$key' is still allowed, but does not have to be.
In the case of the password reset, we are now doing all the controller
logic in the custom password_reset controller method because we need to
handle several round-trips, and we only want to execute on the last one
(but we want to authenticate the pre-auth w/o execute)

A very roughed-in password reset is now fully working, with the only
workflow not yet happening being the e-mail. Right now we just return
the link that we would have sent on teh conf page.

This is checkpoint commit and is still in-progress. The error handling
of this new reworked logic is not yet in place.

IN PROGRESS...