Actor err classes + working fresh phase 2 entry
The phase 2 entry point is now written to handle linking from a fresh
source, such as from the e-mail, and still come back into the
original/origin form using the  local_info system. This relies on saving
the origin referer in the action data

Additionally, error handling is now in place via a error class design in
which the 'request_Actor' method call will always return an 'Actor'
object instance, but now that subclass may be an error object which
cannot be executed but can be seen as an error. And, if it is attempted
to be executed, it simply throws itself.

IN PROGRESS...