Add support for 'error' template (new RapidApp feature)
Now set the newly added RapidApp config option/feature 'error_template'
from the scaffold if 'error' is set and the file exists.

This requires RapidApp 1.3102_03 which just added this feature,
specifically this commit:

https://github.com/vanstyn/RapidApp/commit/32061249fb2c7c1813c

The underlay scaffold does set and have an error template. It is a very
temporary testing file which as of now is based on the same
preauth/login wrapper

Set dev version 1.0200_08