Expand the Action error class + error template
The Action error class now supports additional params including title
and subtitle which are honored by the default error template in the
underlay scaffold. This is able to work safely like this and also still
work when other exceptions are tyhrown. This is based partly on the most
recent RapidApp changes

Additionally the ->throw method can set these new options as a flexible
arg list