EXPERIMENTAL: new PreAuth type 'email_login' - WORKING PROTOTYPE
Added a new preauth action type EmailLogin which simply generates a
one-off link (using the preauth system) to be able to log a user in
without a password. Currently this just displays the link on the screen
for dev purposes, but the idea is that it will e-mail the link, giving
users a quick way to login as long as they have access to the e-mail
address associated with the account. This is no different in level of
security from an ordinary password reset system, except is skips the
step of needing to actually set a new password.

While I do see this as a possibly worthwhile feature, the primary reason
for writing this now is to be able to have a simplier pre-auth use case
to aid in the ongoing development of the feature.

Since this current mode of operation is a total security hole, even
though this commit is never meant to be used in the real world, just to
be safe I've added a die that has to be commented out before the feature
can be used, so there is no commit which exposes a hole like this (and I
want this work to stay in the real master branch)

IN PROGRESS...