Mailer properly handles no recipient case + misc
Removed the hard-coded default_to and added code to actually handle the
situation where there are no recipients (throw exception, but with a
useful message). It would have to be a bug that this code path would get
triggered from a system-generated message, but if that happens, we want
to know about it.

Also added checks to ensure a Mailer object instance only sends once.