Working 'ppRender' API + add TextMarkdown post-processor
Fleshed out the new ppRender template directive which allows manually
passing some arbitrary bit of content through a chosen post-processor.

Up until now, there has been only one post-processor, MarkdownElement
which relies on the marked.js client-side JavaScript library. This works
great, but the downside is any location where this isn't available, such
as in a news feed or a repost situation.

MarkdownElement is still the default, but the new 'TextMarkdown' is now
available as well, and can be accessed via the new ppRender directive.
TextMarkdown uses Text::Markdown, which is a pure pure implementation,
but is not as feature-rich as marked.js, doesn't know about GFM, etc.

The ways ppRender can be called are in source-code comments, still need
to update the POD.

Set dev version 1.1100_01