REFACTOR SCAFFOLD 'PATHS' PARAMS FROM REGEX TO GLOBLIKE
Made the decsion to switch from using regex to using a form of
glob/wildcard matching for the scaffold config params such as
'static_paths'. While this is technically somewhat less powerful, it
makes much more sense and allows common configs to be entered with much
less effort.

During this refactor, a number of bugs were uncovered in the existing
regex implementation anyway, and now a fairly conprehensive unit test is
in place.

I am 99.9% sure that no one out there is using an exotic regex config
that will not continue to work after this change. But, the benefit is
that with this change, patterns like '*.pdf' just work and do what is
expected.

Still need to update docs and probably should also update the built-in
scaffolds.

Set dev version 1.1200_01