ADDED NEW SCAFFOLD CONFIG PARAM 'template_names'
The new optional template config param template_names is the inverse of
static_paths in that it is inclusive of paths that will be treated as
active templates rather than static_paths exclusive behavior. If
template_names is set, all on-disk scaffold files which do not match one
of its entries will be treated as static, regardless of whether or not
they are also covered by static_paths.

This is also slightly different as it applies to effective paths,
post-translation, not the direct public url. For instance, if there is a
default extension of 'html' set, and the path is '/foo' and there is a
file on disk named 'foo.html' in the root of the scaffold, the path
considered for the template_names check will be 'foo.html' not 'foo'.
The same appplies to view wrappers

Still need to update POD, etc

Set dev version 1.1200_02